refactor(frontend): extract chat sub-widgets from WorkspaceChatState (#969)
Break up the 1155-line WorkspaceChatState into focused sub-widgets:
- ChatMessageList: message rendering, scroll/load-more, collapsible messages
- ChatInputBar: text input, @mention autocomplete, emacs keybindings, history
- ChatPresenceBar: connected user avatars
- AgentThinkingIndicator: agent thinking spinner

WorkspaceChatState shrinks to 259 lines (chat state + WS subscriptions),
with build() as a clean composition of sub-widgets. Add 83 new tests
for the extracted widgets.
